Update ClosedXML to version 0.97.0
Updated the ClosedXML package reference in the ClosedXML.SimpleSheets.fsproj file from version 0.96.0 to 0.97.0. This update includes new features, bug fixes, and other improvements provided in the newer version of the ClosedXML library.

One of the improvements is that it removes dependancy on System.Drawing.Common, which was causing a security warning in NuGet audits.