Update ClosedXML, fix typos, and refactor Excel fields
Updated the `ClosedXML` package version from `0.97.0` to `0.102.3` in the `ClosedXML.SimpleSheets.fsproj` file.

Fixed a typo in the module name from `Extenstions` to `Extensions` in the `Excel.fs` file.

Modified the `dateFormat` method in the `FieldMap<'T>` type to remove the line setting `cell.DataType` to `XLDataType.DateTime`.

Refactored the `Excel.field` methods for `int option`, `DateTime option`, `bool option`, and `double option` in the `Excel` type to include intermediate variable assignments for the nullable values.
Specifically for the `bool option` field, added the use of `XLCellValue.FromObject` to set the cell value.
Refactored the `Excel.field` method for `int64 option` to include an intermediate variable assignment for the converted double value.